Continue generator decomposition for v3.4.0
