+++ src/cja_auto_sdr/generator.py
@@ -2033,759 +2033,82 @@ from cja_auto_sdr.diff.snapshot import SnapshotManager, parse_retention_period
 
 # ==================== PROFILE MANAGEMENT ====================
 
+def _profiles_module():
+    from cja_auto_sdr.core import profiles as _profiles
+
+    return _profiles
+
 
 def get_cja_home() -> Path:
-    """Get CJA home directory (~/.cja or $CJA_HOME).
-
-    Returns:
-        Path to CJA home directory
-    """
-    cja_home = os.environ.get("CJA_HOME")
-    if cja_home:
-        return Path(cja_home).expanduser()
-    return Path.home() / ".cja"
+    return _profiles_module().get_cja_home()
 
 
 def get_profiles_dir() -> Path:
-    """Get profiles directory (~/.cja/orgs/).
-
-    Returns:
-        Path to profiles directory
-    """
-    return get_cja_home() / "orgs"
+    return _profiles_module().get_profiles_dir()
 
 
 def get_profile_path(profile_name: str) -> Path:
-    """Get path to a specific profile directory.
-
-    Args:
-        profile_name: Name of the profile
-
-    Returns:
-        Path to profile directory
-    """
-    return get_profiles_dir() / profile_name
+    return _profiles_module().get_profile_path(profile_name)
 
 
 def validate_profile_name(name: str) -> tuple[bool, str | None]:
-    """Validate profile name (alphanumeric, dashes, underscores only).
-
-    Args:
-        name: Profile name to validate
-
-    Returns:
-        Tuple of (is_valid, error_message). error_message is None if valid.
-    """
-    if not name:
-        return False, "Profile name cannot be empty"
-
-    if not re.match(r"^[a-zA-Z0-9][a-zA-Z0-9_-]*$", name):
-        return False, (
-            f"Profile name '{name}' is invalid. "
-            "Must start with alphanumeric and contain only letters, numbers, dashes, and underscores."
-        )
-
-    if len(name) > 64:
-        return False, f"Profile name '{name}' is too long (max 64 characters)"
-
-    return True, None
+    return _profiles_module().validate_profile_name(name)
 
 
 def load_profile_config_json(profile_path: Path) -> dict[str, str] | None:
-    """Load credentials from profile's config.json.
-
-    Args:
-        profile_path: Path to profile directory
-
-    Returns:
-        Dictionary with credentials if config.json exists and is valid, None otherwise
-    """
-    config_file = profile_path / "config.json"
-    if not config_file.exists():
-        return None
-
-    try:
-        with open(config_file) as f:
-            config = json.load(f)
-        if isinstance(config, dict):
-            return filter_credentials(config)
-        return None
-    # PEP 758 (Python 3.14+): `except A, B:` is equivalent to `except (A, B):`.
-    except OSError, json.JSONDecodeError:
-        return None
+    return _profiles_module().load_profile_config_json(profile_path)
 
 
 def load_profile_dotenv(profile_path: Path) -> dict[str, str] | None:
-    """Load credentials from profile's .env file.
-
-    Args:
-        profile_path: Path to profile directory
-
-    Returns:
-        Dictionary with credentials if .env exists and is valid, None otherwise.
-        Malformed/unreadable files are treated as missing.
-    """
-    env_file = profile_path / ".env"
-    if not env_file.exists():
-        return None
-
-    credentials = {}
-    try:
-        # Profiles are written by this tool as UTF-8; decode failures are handled
-        # as unreadable input so one bad file does not break profile operations.
-        with open(env_file, encoding="utf-8") as f:
-            for line in f:
-                line = line.strip()
-                if not line or line.startswith("#"):
-                    continue
-                if "=" in line:
-                    key, _, value = line.partition("=")
-                    key = key.strip()
-                    value = value.strip().strip('"').strip("'")
-                    if key and value:
-                        # Map env var names to config field names
-                        config_key = key.lower()
-                        # Use CREDENTIAL_FIELDS for allowed fields (single source of truth)
-                        if config_key in CREDENTIAL_FIELDS["all"]:
-                            credentials[config_key] = value
-    # PEP 758 (Python 3.14+): `except A, B:` is equivalent to `except (A, B):`.
-    except OSError, UnicodeDecodeError:
-        return None
-
-    filtered_credentials = filter_credentials(credentials)
-    return filtered_credentials or None
+    return _profiles_module().load_profile_dotenv(profile_path)
 
 
 def load_profile_credentials(profile_name: str, logger: logging.Logger) -> dict[str, str] | None:
-    """Load and merge credentials from profile (config.json + .env).
-
-    Precedence within profile:
-    1. .env values override config.json values
-    2. Matches existing behavior (env vars > config file)
-
-    Args:
-        profile_name: Name of the profile to load
-        logger: Logger instance
-
-    Returns:
-        Dictionary with merged credentials, or None if profile not found
-
-    Raises:
-        ProfileNotFoundError: If profile directory doesn't exist
-        ProfileConfigError: If profile has no valid configuration
-    """
-    # Validate profile name
-    is_valid, error_msg = validate_profile_name(profile_name)
-    if not is_valid:
-        raise ProfileConfigError(error_msg, profile_name=profile_name)
-
-    profile_path = get_profile_path(profile_name)
-
-    if not profile_path.exists():
-        raise ProfileNotFoundError(
-            f"Profile '{profile_name}' not found",
-            profile_name=profile_name,
-            details=f"Expected directory: {profile_path}",
-        )
-
-    if not profile_path.is_dir():
-        raise ProfileConfigError(
-            "Profile path is not a directory",
-            profile_name=profile_name,
-            details=str(profile_path),
-        )
-
-    # Load config.json first
-    credentials = load_profile_config_json(profile_path) or {}
-    json_source = bool(credentials)
-
-    # Override with .env values (skip empty values to avoid overwriting valid config)
-    env_credentials = load_profile_dotenv(profile_path)
-    if env_credentials:
-        credentials.update({k: v for k, v in env_credentials.items() if v})
-
-    if not credentials:
-        raise ProfileConfigError(
-            f"Profile '{profile_name}' has no configuration",
-            profile_name=profile_name,
-            details=f"Expected config.json or .env in {profile_path}",
-        )
-
-    if json_source and env_credentials:
-        logger.debug(f"Profile '{profile_name}': merged config.json with .env overrides")
-    elif json_source:
-        logger.debug(f"Profile '{profile_name}': loaded from config.json")
-    else:
-        logger.debug(f"Profile '{profile_name}': loaded from .env")
-
-    return credentials
+    return _profiles_module().load_profile_credentials(profile_name, logger)
 
 
 def resolve_active_profile(cli_profile: str | None = None) -> str | None:
-    """Resolve active profile: --profile > CJA_PROFILE > None.
-
-    Args:
-        cli_profile: Profile name from --profile CLI argument
-
-    Returns:
-        Active profile name, or None if no profile is active
-    """
-    if cli_profile:
-        return cli_profile
-    return os.environ.get("CJA_PROFILE")
+    return _profiles_module().resolve_active_profile(cli_profile)
 
 
 def _read_profile_org_id(profile_path: Path) -> str | None:
-    """Read org_id from a profile directory (best-effort, no validation).
-
-    Matches ``load_profile_credentials`` precedence: .env values override
-    config.json values.  We read config.json first as a base, then let .env
-    override if present.
-
-    Args:
-        profile_path: Path to the profile directory
-
-    Returns:
-        The org_id string, or None if not found or unreadable
-    """
-    org_id: str | None = None
-
-    config_file = profile_path / "config.json"
-    if config_file.exists():
-        try:
-            with open(config_file) as f:
-                data = json.load(f)
-            if isinstance(data, dict):
-                val = data.get("org_id")
-                if val and isinstance(val, str) and val.strip():
-                    org_id = val.strip()
-        # PEP 758 (Python 3.14+): `except A, B:` is equivalent to `except (A, B):`.
-        except OSError, json.JSONDecodeError, ValueError:
-            pass
-
-    # Override: .env (takes precedence, matching load_profile_credentials)
-    env_credentials = load_profile_dotenv(profile_path)
-    if env_credentials:
-        env_org_id = env_credentials.get("org_id")
-        if isinstance(env_org_id, str):
-            normalized_org_id = env_org_id.strip()
-            if normalized_org_id:
-                org_id = normalized_org_id
-
-    return org_id
+    return _profiles_module()._read_profile_org_id(profile_path)
 
 
 def list_profiles(output_format: str = "table") -> bool:
-    """List all profiles with status indicators.
-
-    Args:
-        output_format: Output format - "table" (default) or "json"
-
-    Returns:
-        True if successful, False otherwise
-    """
-    profiles_dir = get_profiles_dir()
-
-    if not profiles_dir.exists():
-        if output_format == "json":
-            print(json.dumps({"profiles": [], "count": 0}, indent=2))
-        else:
-            print()
-            print("No profiles directory found.")
-            print(f"Expected: {profiles_dir}")
-            print()
-            print("To create profiles, run:")
-            print("  cja_auto_sdr --profile-add <profile-name>")
-            print()
-        return True
-
-    profiles = []
-    active_profile = os.environ.get("CJA_PROFILE")
-
-    for item in sorted(profiles_dir.iterdir()):
-        if not item.is_dir():
-            continue
-
-        profile_name = item.name
-        has_config_json = (item / "config.json").exists()
-        has_env = (item / ".env").exists()
-        is_active = profile_name == active_profile
-
-        if has_config_json or has_env:
-            config_source = []
-            if has_config_json:
-                config_source.append("config.json")
-            if has_env:
-                config_source.append(".env")
-
-            try:
-                org_id = _read_profile_org_id(item)
-            except Exception:  # Intentional: Best-effort metadata enrichment; listing must remain available
-                # Best-effort metadata enrichment only; listing must remain
-                # available even if an individual profile cannot be read.
-                org_id = None
-
-            profiles.append(
-                {
-                    "name": profile_name,
-                    "active": is_active,
-                    "sources": config_source,
-                    "path": str(item),
-                    "org_id": org_id,
-                }
-            )
-
-    if output_format == "json":
-        print(json.dumps({"profiles": profiles, "count": len(profiles), "active": active_profile}, indent=2))
-    else:
-        print()
-        print("=" * BANNER_WIDTH)
-        print("AVAILABLE PROFILES")
-        print("=" * BANNER_WIDTH)
-        print()
-
-        if not profiles:
-            print("No profiles found.")
-            print()
-            print("To create a profile, run:")
-            print("  cja_auto_sdr --profile-add <profile-name>")
-        else:
-            # Column widths: marker(2) + name(23) + org_id(32) + sources(20) + status
-            max_org_width = 30
-            print(f"  {'Profile':<23} {'Org ID':<{max_org_width}}  {'Sources':<20} {'Status'}")
-            print("  " + "-" * 60)
-            for p in profiles:
-                status = "[active]" if p["active"] else ""
-                sources = ", ".join(p["sources"])
-                marker = "\u25cf" if p["active"] else " "
-                org_display = p["org_id"] or "\u2014"
-                if len(org_display) > max_org_width:
-                    org_display = org_display[: max_org_width - 1] + "\u2026"
-                print(f"{marker} {p['name']:<23} {org_display:<{max_org_width}}  {sources:<20} {status}")
-
-            print()
-            print(f"Total: {len(profiles)} profile(s)")
-            print()
-            print("Usage:")
-            print("  cja_auto_sdr --profile <name> --list-dataviews")
-            print("  export CJA_PROFILE=<name>")
-
-        print()
-
-    return True
+    return _profiles_module().list_profiles(output_format=output_format)
 
 
 def add_profile_interactive(profile_name: str) -> bool:
-    """Interactively create a new profile.
-
-    Args:
-        profile_name: Name for the new profile
-
-    Returns:
-        True if profile created successfully, False otherwise
-    """
-    # Validate profile name
-    is_valid, error_msg = validate_profile_name(profile_name)
-    if not is_valid:
-        print(ConsoleColors.error(f"Error: {error_msg}"), file=sys.stderr)
-        return False
-
-    profile_path = get_profile_path(profile_name)
-
-    if profile_path.exists():
-        print(f"Profile '{profile_name}' already exists at: {profile_path}")
-        print()
-        response = input("Overwrite? [y/N]: ").strip().lower()
-        if response != "y":
-            print("Aborted.")
-            return False
-
-    # Create profile directory
-    try:
-        profile_path.mkdir(parents=True, exist_ok=True)
-    except OSError as e:
-        print(ConsoleColors.error(f"Error creating profile directory: {e}"), file=sys.stderr)
-        return False
-
-    print()
-    print("=" * BANNER_WIDTH)
-    print(f"CREATING PROFILE: {profile_name}")
-    print("=" * BANNER_WIDTH)
-    print()
-    print("Enter your Adobe OAuth credentials.")
-    print("(Get these from Adobe Developer Console → Project → Credentials)")
-    print()
-
-    try:
-        org_id = input("Organization ID (ends with @AdobeOrg): ").strip()
-        if not org_id:
-            print(ConsoleColors.error("Error: Organization ID is required"), file=sys.stderr)
-            return False
-
-        client_id = input("Client ID: ").strip()
-        if not client_id:
-            print(ConsoleColors.error("Error: Client ID is required"), file=sys.stderr)
-            return False
-
-        # Use getpass for secret (never fall back to echoing input)
-        import getpass
-
-        try:
-            secret = getpass.getpass("Client Secret: ").strip()
-        except getpass.GetPassWarning:
-            print(
-                ConsoleColors.error(
-                    "Error: Cannot securely read secret (no TTY available). Use --profile-import instead."
-                ),
-                file=sys.stderr,
-            )
-            return False
-
-        if not secret:
-            print(ConsoleColors.error("Error: Client Secret is required"), file=sys.stderr)
-            return False
-
-        scopes = input("OAuth Scopes (from Developer Console): ").strip()
-        if not scopes:
-            print(ConsoleColors.error("Error: OAuth Scopes are required"), file=sys.stderr)
-            return False
-
-    except KeyboardInterrupt, EOFError:
-        print("\nAborted.")
-        return False
-
-    # Create config.json
-    config = {"org_id": org_id, "client_id": client_id, "secret": secret, "scopes": scopes}
-
-    config_file = profile_path / "config.json"
-    try:
-        fd = os.open(str(config_file), os.O_WRONLY | os.O_CREAT | os.O_TRUNC, 0o600)
-        with open(fd, "w", encoding="utf-8") as f:
-            json.dump(config, f, indent=2)
-    except OSError as e:
-        print(ConsoleColors.error(f"Error writing config file: {e}"), file=sys.stderr)
-        return False
-
-    print()
-    print(f"Profile '{profile_name}' created successfully!")
-    print(f"  Location: {profile_path}")
-    print()
-    print("Test your profile:")
-    print(f"  cja_auto_sdr --profile-test {profile_name}")
-    print()
-    print("Use your profile:")
-    print(f"  cja_auto_sdr --profile {profile_name} --list-dataviews")
-    print()
-
-    return True
+    return _profiles_module().add_profile_interactive(profile_name)
 
 
 def _normalize_import_credentials(raw_credentials: dict[str, Any]) -> dict[str, str]:
-    """Normalize imported credential keys/values to canonical config field names."""
-    env_to_config_key = {env_name.lower(): config_key for config_key, env_name in ENV_VAR_MAPPING.items()}
-    alias_map = {
-        "organization_id": "org_id",
-        "orgid": "org_id",
-        "clientid": "client_id",
-        "client_secret": "secret",
-        "clientsecret": "secret",
-    }
-
-    normalized: dict[str, str] = {}
-    for raw_key, raw_value in raw_credentials.items():
-        if raw_value is None:
-            continue
-        key = _canonical_quality_policy_key(raw_key)
-        key = env_to_config_key.get(key, key)
-        key = alias_map.get(key, key)
-        if key not in CREDENTIAL_FIELDS["all"]:
-            continue
-        value = str(raw_value).strip().strip('"').strip("'")
-        if value:
-            normalized[key] = value
-
-    return normalized
+    return _profiles_module()._normalize_import_credentials(raw_credentials)
 
 
 def _parse_env_credentials_content(content: str) -> dict[str, str]:
-    """Parse .env-formatted credentials content."""
-    credentials: dict[str, str] = {}
-    for line_number, raw_line in enumerate(content.splitlines(), start=1):
-        line = raw_line.strip()
-        if not line or line.startswith("#"):
-            continue
-        if line.lower().startswith("export "):
-            line = line[7:].strip()
-        if "=" not in line:
-            raise ValueError(f"Invalid .env content at line {line_number}: expected KEY=VALUE")
-        key, _, value = line.partition("=")
-        key = key.strip()
-        value = value.strip()
-        if not key:
-            raise ValueError(f"Invalid .env content at line {line_number}: empty key")
-        normalized = _normalize_import_credentials({key: value})
-        credentials.update(normalized)
-
-    return credentials
+    return _profiles_module()._parse_env_credentials_content(content)
 
 
 def load_profile_import_source(source_file: str | Path) -> dict[str, str]:
-    """Load credentials from a JSON/.env file or from a directory containing profile files."""
-    source_path = Path(source_file).expanduser()
-    if not source_path.exists():
-        raise FileNotFoundError(f"Source not found: {source_path}")
-
-    if source_path.is_dir():
-        config_json_path = source_path / "config.json"
-        dotenv_path = source_path / ".env"
-        merged_credentials: dict[str, str] = {}
-
-        if config_json_path.exists():
-            with open(config_json_path, encoding="utf-8") as f:
-                config_payload = json.load(f)
-            if not isinstance(config_payload, dict):
-                raise ValueError(f"Invalid profile config in {config_json_path}: expected JSON object")
-            if isinstance(config_payload.get("credentials"), dict):
-                config_payload = config_payload["credentials"]
-            merged_credentials.update(_normalize_import_credentials(config_payload))
-
-        if dotenv_path.exists():
-            env_credentials = _parse_env_credentials_content(dotenv_path.read_text(encoding="utf-8"))
-            merged_credentials.update(env_credentials)
-
-        if not merged_credentials:
-            raise ValueError(f"No credentials found in directory: {source_path}")
-        return merged_credentials
-
-    if source_path.suffix.lower() == ".json":
-        with open(source_path, encoding="utf-8") as f:
-            payload = json.load(f)
-        if not isinstance(payload, dict):
-            raise ValueError("Profile import JSON must be an object")
-        if isinstance(payload.get("credentials"), dict):
-            payload = payload["credentials"]
-        credentials = _normalize_import_credentials(payload)
-    else:
-        credentials = _parse_env_credentials_content(source_path.read_text(encoding="utf-8"))
-
-    if not credentials:
-        raise ValueError("No supported credential fields were found in import source")
-    return credentials
+    return _profiles_module().load_profile_import_source(source_file)
 
 
 def import_profile_non_interactive(profile_name: str, source_file: str | Path, overwrite: bool = False) -> bool:
-    """Import a profile from JSON/.env without interactive prompts."""
-    is_valid_name, error_msg = validate_profile_name(profile_name)
-    if not is_valid_name:
-        print(ConsoleColors.error(f"Error: {error_msg}"), file=sys.stderr)
-        return False
-
-    profile_path = get_profile_path(profile_name)
-    profile_exists = profile_path.exists()
-    if profile_exists and not overwrite:
-        print(f"Profile '{profile_name}' already exists at: {profile_path}")
-        print("Use --profile-overwrite with --profile-import to replace it.")
-        return False
-
-    try:
-        credentials = load_profile_import_source(source_file)
-    except (OSError, json.JSONDecodeError, ValueError) as e:
-        print(ConsoleColors.error(f"Error loading profile import source '{source_file}': {e}"), file=sys.stderr)
-        return False
-
-    validation_logger = logging.getLogger("profile_import")
-    validation_logger.setLevel(logging.WARNING)
-    is_usable, issues = validate_credentials(
-        credentials,
-        validation_logger,
-        strict=True,
-        source=f"profile import ({source_file})",
-    )
-    if not is_usable:
-        _print_error_list_to_stderr("Error: Imported credentials failed validation:", issues)
-        return False
-
-    config_payload = {
-        key: value for key in ("org_id", "client_id", "secret", "scopes", "sandbox") if (value := credentials.get(key))
-    }
-
-    try:
-        profile_path.mkdir(parents=True, exist_ok=True)
-        config_path = profile_path / "config.json"
-        fd = os.open(str(config_path), os.O_WRONLY | os.O_CREAT | os.O_TRUNC, 0o600)
-        with open(fd, "w", encoding="utf-8") as f:
-            json.dump(config_payload, f, indent=2)
-            f.write("\n")
-    except OSError as e:
-        print(ConsoleColors.error(f"Error writing profile config: {e}"), file=sys.stderr)
-        return False
-
-    if profile_exists and (profile_path / ".env").exists():
-        print(
-            "Warning: Existing .env file was kept and may override imported config.json values when this profile is used.",
-        )
-
-    print()
-    print(f"Profile '{profile_name}' imported successfully.")
-    print(f"  Source: {Path(source_file).expanduser()}")
-    print(f"  Location: {profile_path}")
-    print()
-    print("Next steps:")
-    print(f"  cja_auto_sdr --profile-test {profile_name}")
-    print(f"  cja_auto_sdr --profile {profile_name} --list-dataviews")
-    print()
-    return True
+    return _profiles_module().import_profile_non_interactive(profile_name, source_file, overwrite=overwrite)
 
 
 def mask_sensitive_value(value: str, show_chars: int = 4) -> str:
-    """Mask sensitive values for display.
-
-    Args:
-        value: The value to mask
-        show_chars: Number of characters to show at start and end
-
-    Returns:
-        Masked value string
-    """
-    if not value:
-        return "(empty)"
-
-    if len(value) <= show_chars * 2:
-        return "*" * len(value)
-
-    return f"{value[:show_chars]}{'*' * (len(value) - show_chars * 2)}{value[-show_chars:]}"
+    return _profiles_module().mask_sensitive_value(value, show_chars=show_chars)
 
 
 def show_profile(profile_name: str) -> bool:
-    """Display profile config with masked secrets.
-
-    Args:
-        profile_name: Name of the profile to show
-
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        logger = logging.getLogger("profile_show")
-        logger.setLevel(logging.WARNING)
-        credentials = load_profile_credentials(profile_name, logger)
-    except ProfileNotFoundError as e:
-        print(ConsoleColors.error(f"Error: {e}"), file=sys.stderr)
-        return False
-    except ProfileConfigError as e:
-        print(ConsoleColors.error(f"Error: {e}"), file=sys.stderr)
-        return False
-
-    profile_path = get_profile_path(profile_name)
-
-    print()
-    print("=" * BANNER_WIDTH)
-    print(f"PROFILE: {profile_name}")
-    print("=" * BANNER_WIDTH)
-    print()
-    print(f"Location: {profile_path}")
-    print()
-
-    # Show sources
-    sources = []
-    if (profile_path / "config.json").exists():
-        sources.append("config.json")
-    if (profile_path / ".env").exists():
-        sources.append(".env")
-    print(f"Sources: {', '.join(sources)}")
-    print()
-
-    # Show credentials with masked values
-    print("Credentials:")
-    print("-" * 40)
-
-    sensitive_fields = {"secret", "client_id"}
-
-    for key in ["org_id", "client_id", "secret", "scopes", "sandbox"]:
-        if key in credentials:
-            value = credentials[key]
-            if key in sensitive_fields:
-                display_value = mask_sensitive_value(value)
-            else:
-                display_value = value
-            print(f"  {key}: {display_value}")
-
-    print()
-    return True
+    return _profiles_module().show_profile(profile_name)
 
 
 def test_profile(profile_name: str) -> bool:
-    """Test profile credentials and API connectivity.
-
-    Args:
-        profile_name: Name of the profile to test
-
-    Returns:
-        True if test successful, False otherwise
-    """
-    print()
-    print("=" * BANNER_WIDTH)
-    print(f"TESTING PROFILE: {profile_name}")
-    print("=" * BANNER_WIDTH)
-    print()
-
-    # Load credentials
-    try:
-        logger = logging.getLogger("profile_test")
-        logger.setLevel(logging.WARNING)
-        credentials = load_profile_credentials(profile_name, logger)
-    except ProfileNotFoundError as e:
-        print(ConsoleColors.error(f"ERROR: {e}"), file=sys.stderr)
-        return False
-    except ProfileConfigError as e:
-        print(ConsoleColors.error(f"ERROR: {e}"), file=sys.stderr)
-        return False
-
-    print("1. Profile found and loaded")
-
-    # Validate credentials
-    issues = ConfigValidator.validate_all(credentials, logger)
-    if issues:
-        print("2. Credential validation: WARNINGS")
-        for issue in issues:
-            print(f"   - {issue}")
-    else:
-        print("2. Credential validation: OK")
-
-    # Test API connectivity
-    print("3. Testing API connectivity...")
-
-    try:
-        _config_from_env(credentials, logger)
-        cja = cjapy.CJA()
-        dataviews = cja.getDataViews()
-
-        if dataviews is not None:
-            count = len(dataviews) if hasattr(dataviews, "__len__") else 0
-            print("   API connection: SUCCESS")
-            print(f"   Data views accessible: {count}")
-            print()
-            print("Profile test: PASSED")
-            print()
-            return True
-        print("   API connection: OK (no data views found)")
-        print()
-        print("Profile test: PASSED")
-        print()
-        return True
-
-    except RECOVERABLE_CONFIG_API_EXCEPTIONS as e:  # cjapy API/bootstrap calls
-        print(ConsoleColors.error("   API connection: FAILED"), file=sys.stderr)
-        print(ConsoleColors.error(f"   Error: {e}"), file=sys.stderr)
-        print()
-        print(ConsoleColors.error("Profile test: FAILED"), file=sys.stderr)
-        print()
-        print("Common issues:")
-        print("  - Invalid client_id or secret")
-        print("  - Incorrect OAuth scopes")
-        print("  - API project not provisioned for CJA")
-        print()
-        return False
+    return _profiles_module().test_profile(profile_name)
 
 
 # ==================== CONFIG VALIDATION (moved to core/config_validation.py) ====================
@@ -7184,46 +6507,9 @@ def get_cached_data_views(cja, cache_key: str, logger: logging.Logger) -> list[d
 
 
 def prompt_for_selection(options: list[tuple[str, str]], prompt_text: str) -> str | None:
-    """
-    Prompt user to select from a list of options interactively.
+    from cja_auto_sdr.cli.interactive import prompt_for_selection as _impl
 
-    Args:
-        options: List of (id, display_text) tuples
-        prompt_text: Text to display before options
-
-    Returns:
-        Selected ID or None if user cancels
-    """
-    # Check if we're in an interactive terminal
-    if not sys.stdin.isatty():
-        return None
-
-    print(f"\n{prompt_text}")
-    print("-" * 40)
-
-    for i, (opt_id, display) in enumerate(options, 1):
-        print(f"  [{i}] {display}")
-        print(f"      ID: {opt_id}")
-
-    print("  [0] Cancel")
-    print()
-
-    while True:
-        try:
-            choice = input("Enter selection (number): ").strip()
-            if choice == "0" or choice.lower() in ("q", "quit", "cancel"):
-                return None
-
-            idx = int(choice)
-            if 1 <= idx <= len(options):
-                return options[idx - 1][0]
-
-            print(f"Invalid selection. Enter 1-{len(options)} or 0 to cancel.")
-        except ValueError:
-            print("Please enter a number.")
-        except EOFError, KeyboardInterrupt:
-            print("\nCancelled.")
-            return None
+    return _impl(options, prompt_text)
 
 
 @dataclass(frozen=True)
@@ -9406,191 +8692,9 @@ def list_datasets(
 
 
 def interactive_select_dataviews(config_file: str = "config.json", profile: str | None = None) -> list[str]:
-    """
-    Interactively select data views from a numbered list.
+    from cja_auto_sdr.cli.interactive import interactive_select_dataviews as _impl
 
-    Supports selection formats:
-    - Single number: "3"
-    - Multiple numbers: "1,3,5"
-    - Range: "1-5"
-    - Combined: "1,3-5,7"
-    - All: "all" or "a"
-
-    Args:
-        config_file: Path to CJA configuration file
-        profile: Optional profile name to use for credentials
-
-    Returns:
-        List of selected data view IDs, or empty list on error/cancel
-    """
-    print()
-    print("=" * BANNER_WIDTH)
-    print("INTERACTIVE DATA VIEW SELECTION")
-    print("=" * BANNER_WIDTH)
-    print()
-    if profile:
-        print(f"Using profile: {profile}")
-    else:
-        print(f"Using configuration: {config_file}")
-    print()
-
-    try:
-        success, source, _ = configure_cjapy(profile, config_file)
-        if not success:
-            print(ConsoleColors.error(f"ERROR: {source}"))
-            return []
-        cja = cjapy.CJA()
-
-        print("Fetching available data views...")
-        available_dvs = cja.getDataViews()
-
-        if available_dvs is None or (hasattr(available_dvs, "__len__") and len(available_dvs) == 0):
-            print()
-            print(ConsoleColors.warning("No data views found or no access to any data views."))
-            return []
-
-        # Convert to list if DataFrame
-        if isinstance(available_dvs, pd.DataFrame):
-            available_dvs = available_dvs.to_dict("records")
-
-        # Build display data
-        display_data = []
-        for dv in available_dvs:
-            if isinstance(dv, dict):
-                dv_id = _normalize_optional_text(dv.get("id"), default="N/A")
-                dv_name = _normalize_optional_text(dv.get("name"), default="N/A")
-                owner_name = _extract_owner_name(dv.get("owner"))
-                display_data.append({"id": dv_id, "name": dv_name, "owner": owner_name})
-
-        if not display_data:
-            print(ConsoleColors.warning("No data views available."))
-            return []
-
-        # Calculate column widths
-        num_width = len(str(len(display_data))) + 2
-        max_id_width = max(len("ID"), *(len(item["id"]) for item in display_data)) + 2
-        max_name_width = max(len("Name"), *(len(item["name"]) for item in display_data)) + 2
-        max_owner_width = max(len("Owner"), *(len(item["owner"]) for item in display_data)) + 2
-
-        total_width = num_width + max_id_width + max_name_width + max_owner_width
-
-        print()
-        print(f"Found {len(display_data)} accessible data view(s):")
-        print()
-        print(f"{'#':<{num_width}} {'ID':<{max_id_width}} {'Name':<{max_name_width}} {'Owner':<{max_owner_width}}")
-        print("-" * total_width)
-
-        for idx, item in enumerate(display_data, 1):
-            print(
-                f"{idx:<{num_width}} {item['id']:<{max_id_width}} {item['name']:<{max_name_width}} {item['owner']:<{max_owner_width}}",
-            )
-
-        print()
-        print("-" * total_width)
-        print("Selection options:")
-        print("  Single:   3         (selects #3)")
-        print("  Multiple: 1,3,5     (selects #1, #3, #5)")
-        print("  Range:    1-5       (selects #1 through #5)")
-        print("  Combined: 1,3-5,7   (selects #1, #3, #4, #5, #7)")
-        print("  All:      all or a  (selects all data views)")
-        print("  Cancel:   q or quit (exit without selection)")
-        print()
-
-        while True:
-            try:
-                selection = input("Enter selection: ").strip().lower()
-            except EOFError:
-                print()
-                print(ConsoleColors.warning("No input available (non-interactive terminal)."))
-                return []
-
-            if not selection:
-                print("Please enter a selection.")
-                continue
-
-            if selection in ("q", "quit", "exit", "cancel"):
-                print(ConsoleColors.warning("Selection cancelled."))
-                return []
-
-            # Parse selection
-            selected_indices = set()
-
-            if selection in ("all", "a", "*"):
-                selected_indices = set(range(1, len(display_data) + 1))
-            else:
-                # Parse comma-separated parts
-                parts = selection.replace(" ", "").split(",")
-                valid = True
-                for part in parts:
-                    if not part:
-                        continue
-                    if "-" in part:
-                        # Range like "1-5"
-                        try:
-                            range_parts = part.split("-")
-                            if len(range_parts) != 2:
-                                raise ValueError("Invalid range format")
-                            start = int(range_parts[0])
-                            end = int(range_parts[1])
-                            if start > end:
-                                start, end = end, start
-                            for i in range(start, end + 1):
-                                selected_indices.add(i)
-                        except ValueError:
-                            print(ConsoleColors.error(f"Invalid range: '{part}'. Use format like '1-5'."))
-                            valid = False
-                            break
-                    else:
-                        # Single number
-                        try:
-                            num = int(part)
-                            selected_indices.add(num)
-                        except ValueError:
-                            print(ConsoleColors.error(f"Invalid number: '{part}'."))
-                            valid = False
-                            break
-
-                if not valid:
-                    continue
-
-            # Validate indices
-            invalid_indices = [i for i in selected_indices if i < 1 or i > len(display_data)]
-            if invalid_indices:
-                print(
-                    ConsoleColors.error(f"Invalid selection(s): {invalid_indices}. Valid range: 1-{len(display_data)}"),
-                )
-                continue
-
-            if not selected_indices:
-                print("No valid selections. Please try again.")
-                continue
-
-            # Convert indices to IDs
-            selected_ids = [display_data[i - 1]["id"] for i in sorted(selected_indices)]
-
-            print()
-            print(f"Selected {len(selected_ids)} data view(s):")
-            for idx in sorted(selected_indices):
-                item = display_data[idx - 1]
-                print(f"  {idx}. {item['name']} ({item['id']})")
-
-            return selected_ids
-
-    except FileNotFoundError:
-        print(ConsoleColors.error(f"ERROR: Configuration file '{config_file}' not found"))
-        print()
-        print("Generate a sample configuration file with:")
-        print("  cja_auto_sdr --sample-config")
-        return []
-
-    except KeyboardInterrupt, SystemExit:
-        print()
-        print(ConsoleColors.warning("Operation cancelled."))
-        return []
-
-    except RECOVERABLE_COMMAND_HANDLER_EXCEPTIONS as e:
-        print(ConsoleColors.error(f"ERROR: Failed to connect to CJA API: {e!s}"))
-        return []
+    return _impl(config_file=config_file, profile=profile)
 
 
 # ==================== INTERACTIVE MODE ====================
@@ -9610,296 +8714,9 @@ class WizardConfig:
 
 
 def interactive_wizard(config_file: str = "config.json", profile: str | None = None) -> WizardConfig | None:
-    """
-    Interactive wizard for guided SDR generation.
+    from cja_auto_sdr.cli.interactive import interactive_wizard as _impl
 
-    Walks users through:
-    1. Data view selection
-    2. Output format selection
-    3. Inventory options
-    4. Summary and confirmation
-
-    Args:
-        config_file: Path to CJA configuration file
-        profile: Optional profile name to use for credentials
-
-    Returns:
-        WizardConfig with user selections, or None if cancelled
-    """
-
-    def prompt_choice(prompt: str, options: list[tuple[str, str]], default: str | None = None) -> str | None:
-        """Prompt user to select from numbered options."""
-        print()
-        print(prompt)
-        print()
-        for i, (key, label) in enumerate(options, 1):
-            default_marker = " (default)" if key == default else ""
-            print(f"  {i}. {label}{default_marker}")
-        print()
-        print("  q. Cancel and exit")
-        print()
-
-        while True:
-            try:
-                default_hint = f" [{options[[k for k, _ in options].index(default)][1]}]" if default else ""
-                choice = input(f"Enter choice (1-{len(options)}){default_hint}: ").strip().lower()
-            except EOFError, KeyboardInterrupt:
-                print()
-                return None
-
-            if choice in ("q", "quit", "exit", "cancel"):
-                return None
-
-            if not choice and default:
-                return default
-
-            try:
-                idx = int(choice) - 1
-                if 0 <= idx < len(options):
-                    return options[idx][0]
-                print(f"Please enter a number between 1 and {len(options)}.")
-            except ValueError:
-                print(f"Please enter a number between 1 and {len(options)}, or 'q' to cancel.")
-
-    def prompt_yes_no(prompt: str, default: bool = False) -> bool | None:
-        """Prompt user for yes/no answer."""
-        if default:
-            prompt_hint = "[Y/n] (Enter=yes)"
-        else:
-            prompt_hint = "[y/N] (Enter=no)"
-        valid_yes = ("y", "yes", "1", "true")
-        valid_no = ("n", "no", "0", "false")
-        valid_quit = ("q", "quit", "exit", "cancel")
-
-        while True:
-            print()
-            try:
-                answer = input(f"{prompt} {prompt_hint}: ").strip().lower()
-            except EOFError, KeyboardInterrupt:
-                print()
-                return None
-
-            if answer in valid_quit:
-                return None
-            if not answer:
-                return default
-            if answer in valid_yes:
-                return True
-            if answer in valid_no:
-                return False
-
-            # Invalid input - show error and retry
-            print(ConsoleColors.warning(f"Invalid input '{answer}'. Please enter 'y' or 'n' (or 'q' to quit)."))
-
-    print()
-    print("=" * BANNER_WIDTH)
-    print("  CJA SDR GENERATOR - INTERACTIVE MODE")
-    print("=" * BANNER_WIDTH)
-    print()
-    print("This interactive mode will guide you through generating an SDR.")
-    print("Press 'q' at any prompt to cancel.")
-
-    # Step 1: Connect and show data views
-    print()
-    print("-" * 60)
-    print("STEP 1: Select Data View(s)")
-    print("-" * 60)
-
-    if profile:
-        print(f"Using profile: {profile}")
-    else:
-        print(f"Using configuration: {config_file}")
-    print()
-
-    try:
-        success, source, _ = configure_cjapy(profile, config_file)
-        if not success:
-            print(ConsoleColors.error(f"ERROR: {source}"))
-            return None
-        cja = cjapy.CJA()
-
-        print("Fetching available data views...")
-        available_dvs = cja.getDataViews()
-
-        if available_dvs is None or (hasattr(available_dvs, "__len__") and len(available_dvs) == 0):
-            print()
-            print(ConsoleColors.warning("No data views found or no access to any data views."))
-            return None
-
-        # Convert to list if DataFrame
-        if isinstance(available_dvs, pd.DataFrame):
-            available_dvs = available_dvs.to_dict("records")
-
-        # Build display data
-        display_data = []
-        for dv in available_dvs:
-            if isinstance(dv, dict):
-                dv_id = _normalize_optional_text(dv.get("id"), default="N/A")
-                dv_name = _normalize_optional_text(dv.get("name"), default="N/A")
-                display_data.append({"id": dv_id, "name": dv_name})
-
-        if not display_data:
-            print(ConsoleColors.warning("No data views available."))
-            return None
-
-        # Show data views
-        print()
-        print(f"Found {len(display_data)} accessible data view(s):")
-        print()
-        for idx, item in enumerate(display_data, 1):
-            print(f"  {idx}. {item['name']}")
-            print(f"      {ConsoleColors.dim(item['id'])}")
-
-        print()
-        print("Selection options: single (3), multiple (1,3,5), range (1-3), all")
-        print()
-
-        while True:
-            try:
-                selection = input("Select data view(s): ").strip().lower()
-            except EOFError, KeyboardInterrupt:
-                print()
-                return None
-
-            if selection in ("q", "quit", "exit", "cancel"):
-                print(ConsoleColors.warning("Cancelled."))
-                return None
-
-            if not selection:
-                print("Please enter a selection.")
-                continue
-
-            # Parse selection (reuse logic from interactive_select_dataviews)
-            selected_indices = set()
-            valid = True
-
-            if selection in ("all", "a", "*"):
-                selected_indices = set(range(1, len(display_data) + 1))
-            else:
-                parts = selection.replace(" ", "").split(",")
-                for part in parts:
-                    if not part:
-                        continue
-                    if "-" in part:
-                        try:
-                            range_parts = part.split("-")
-                            if len(range_parts) != 2:
-                                raise ValueError
-                            start, end = int(range_parts[0]), int(range_parts[1])
-                            if start > end:
-                                start, end = end, start
-                            for i in range(start, end + 1):
-                                selected_indices.add(i)
-                        except ValueError:
-                            print(ConsoleColors.error(f"Invalid range: '{part}'"))
-                            valid = False
-                            break
-                    else:
-                        try:
-                            selected_indices.add(int(part))
-                        except ValueError:
-                            print(ConsoleColors.error(f"Invalid number: '{part}'"))
-                            valid = False
-                            break
-
-            if not valid:
-                continue
-
-            # Validate
-            invalid = [i for i in selected_indices if i < 1 or i > len(display_data)]
-            if invalid:
-                print(ConsoleColors.error(f"Invalid: {invalid}. Valid range: 1-{len(display_data)}"))
-                continue
-
-            if not selected_indices:
-                continue
-
-            selected_ids = [display_data[i - 1]["id"] for i in sorted(selected_indices)]
-            selected_names = [display_data[i - 1]["name"] for i in sorted(selected_indices)]
-
-            print()
-            print(f"Selected: {', '.join(selected_names)}")
-            break
-
-    except FileNotFoundError:
-        print(ConsoleColors.error(f"ERROR: Configuration file '{config_file}' not found"))
-        print("Run: cja_auto_sdr --sample-config")
-        return None
-    except RECOVERABLE_COMMAND_HANDLER_EXCEPTIONS as e:
-        print(ConsoleColors.error(f"ERROR: Failed to connect to CJA API: {e!s}"))
-        return None
-
-    # Step 2: Output format
-    print()
-    print("-" * 60)
-    print("STEP 2: Choose Output Format")
-    print("-" * 60)
-
-    format_options = [
-        ("excel", "Excel (.xlsx) - Best for review and sharing"),
-        ("json", "JSON - Best for automation and APIs"),
-        ("csv", "CSV - Best for data processing"),
-        ("html", "HTML - Best for web viewing"),
-        ("markdown", "Markdown - Best for documentation/GitHub"),
-        ("all", "All formats - Generate everything"),
-    ]
-
-    output_format = prompt_choice("Which output format would you like?", format_options, default="excel")
-    if output_format is None:
-        print(ConsoleColors.warning("Cancelled."))
-        return None
-
-    # Step 3: Inventory options
-    print()
-    print("-" * 60)
-    print("STEP 3: Include Inventory Data?")
-    print("-" * 60)
-    print()
-    print("Inventory data provides additional documentation beyond the standard SDR:")
-    print("  • Segments: Filter definitions, complexity scores, references")
-    print("  • Calculated Metrics: Formulas, complexity, metric dependencies")
-    print("  • Derived Fields: Logic analysis, functions used, schema references")
-
-    include_segments = prompt_yes_no("Include Segments inventory?", default=False)
-    if include_segments is None:
-        print(ConsoleColors.warning("Cancelled."))
-        return None
-
-    include_calculated = prompt_yes_no("Include Calculated Metrics inventory?", default=False)
-    if include_calculated is None:
-        print(ConsoleColors.warning("Cancelled."))
-        return None
-
-    include_derived = prompt_yes_no("Include Derived Fields inventory?", default=False)
-    if include_derived is None:
-        print(ConsoleColors.warning("Cancelled."))
-        return None
-
-    # Step 4: Summary and confirmation
-    print()
-    print("-" * 60)
-    print("SUMMARY")
-    print("-" * 60)
-    print()
-    print(f"  Data View(s):        {', '.join(selected_names)}")
-    print(f"  Output Format:       {output_format.upper()}")
-    print(f"  Include Segments:    {'Yes' if include_segments else 'No'}")
-    print(f"  Include Calc Metrics: {'Yes' if include_calculated else 'No'}")
-    print(f"  Include Derived:     {'Yes' if include_derived else 'No'}")
-    print()
-
-    confirm = prompt_yes_no("Generate SDR with these settings?", default=True)
-    if not confirm:
-        print(ConsoleColors.warning("Cancelled."))
-        return None
-
-    return WizardConfig(
-        data_view_ids=selected_ids,
-        output_format=output_format,
-        include_segments=include_segments,
-        include_calculated=include_calculated,
-        include_derived=include_derived,
-    )
+    return _impl(config_file=config_file, profile=profile)
 
 
 # ==================== SAMPLE CONFIG GENERATOR ====================
